OmicsKit: A bioinformatics toolset for multiomics analysis

This package contains functions that help in manipulating tables and generating plots for multi-omics analysis including genomics, transcriptomics, proteomics, methylomics and immunoinformatics.

Author

Maintainer: Daniel Guevara dgdiaz011202@gmail.com (ORCID)

Authors:

Daniel Garbozo dan.garbozo.1201.urp@gmail.com (ORCID)

Angela Alarcon alarcondcg@gmail.com (ORCID)

Jennifer Almeyda jalmeydatz@gmail.com (ORCID)

David Requena david.requena@nyulangone.org (ORCID)